New minecraft version checker
this exe has two main files. a python file and a nodejs file. the nodejs file opens an api which it gets the minecraft versions into and all that type stuff, and the python file makes a txt which it saves all the versions into, and if theres a new version, it sends a notification (toast) to windows. if the notif is clicked, it links to the minecraft.net page of that version. I do not plan updating this, it already works fine